refactor compra y proveedor

+++ Features/Report/DTO/ProductoVendidoDTO.cs
@@ -5,7 +5,7 @@ namespace proyecto_venta_stock.Report.DTO
         public int IdProducto { get; set; }
         public string NombreProducto { get; set; } = null!;
         public string Categoria { get; set; } = null!;
-        public int CantidadVendida { get; set; }
+        public decimal CantidadVendida { get; set; }
         public decimal TotalFacturado { get; set; }
     }
 }
