+++ Models/CompraProveedorDetalle.cs
@@ -8,7 +8,7 @@ public partial class CompraProveedorDetalle
 
     public int IdProducto { get; set; }
 
-    public int Cantidad { get; set; }
+    public decimal Cantidad { get; set; }
 
     public decimal PrecioUnitario { get; set; }
